test(tp): move jax transforms tests, parameterize test classes
