+++ tests/test_ops/test_tensor_product_op.py
@@ -12,15 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import re
-
 import jax
 import jax.numpy as np
 import jax.random as random
 import numpy.testing as testing
 import pytest
 from jax import Array
-from jax.experimental.topologies import get_topology_desc
 
 import e3j
 from e3j.ops import TensorProductParams as Params
@@ -32,10 +29,20 @@ e3j.config(
     debug_level=0,
 )
 
-NUM_STRIPS = 10
-LEN_STRIPS = 6
-NUM_ROWS = 128
-NUM_CHANNELS = 32
+
+@pytest.fixture(scope="class", params=[32, 64, 128, 512])
+def channels(request):
+    request.cls.channels_x = request.param
+    if getattr(request.cls, "mode", None) in ("INNER", "MAP"):
+        request.cls.channels_y = request.param
+    elif getattr(request.cls, "mode", None) == "OUTER":
+        request.cls.channels_y = None
+
+
+@pytest.fixture(scope="class", params=["OUTER", "INNER", "MAP"])
+def mode(request):
+    request.cls.mode = request.param
+
 
 np.set_printoptions(
     precision=4,
@@ -63,42 +70,6 @@ def assert_allclose(expect, result, rtol=5e-6, atol=5e-6, debug: int = 1):
         raise err
 
 
-def generate_tp_data(
-    channels: tuple[int | None, int | None] = (None, None),
-) -> tuple[int, Array, Array, Array, Array]:
-    """
-    Generate common test inputs for the forward and backward tests.
-
-    Returns:
-        num_out: target space dimension.
-        idx: A 2D index array of leading dimension 3.
-        coef: A 1D value vector of length `idx.shape[-1]`.
-        x: A 2D array of shape `(..., num_x)`
-        y: A 2D array of shape `(..., num_y)`
-    """
-    num_rows = NUM_ROWS
-
-    # lexsorted indices
-    num_out, num_x, num_y = 2, 2, 3
-    idx_0 = np.tile(np.array([0, 0, 0, 1, 1, 1]), NUM_STRIPS)
-    idx_1 = np.tile(np.array([0, 0, 0, 1, 1, 1]), NUM_STRIPS)
-    idx_2 = np.tile(np.array([0, 1, 2, 0, 1, 2]), NUM_STRIPS)
-    offsets = np.array([num_out, num_x, num_y])[:, None]
-    offsets *= np.arange(NUM_STRIPS).repeat(LEN_STRIPS)
-    idx = np.stack([idx_0, idx_1, idx_2]) + offsets
-
-    x = np.stack([np.tile(np.array([1.0, 2.0]), NUM_STRIPS)] * num_rows)
-    y = np.stack([np.tile(np.array([3.0, 4.0, 5.0]), NUM_STRIPS)] * num_rows)
-
-    if channels[0] is not None:
-        x = np.tile(x[:, None, :], (1, channels[0], 1))
-    if channels[1] is not None:
-        y = np.tile(y[:, None, :], (1, channels[1], 1))
-
-    coef = np.ones(idx.shape[-1])
-    return num_out * NUM_STRIPS, idx, coef, x, y
-
-
 def tensor_product_reference(
     idx, coef, x, y, num_out, mode="OUTER", layout="LEADING_CHANNELS"
 ):
@@ -115,6 +86,10 @@ def tensor_product_reference(
     n_rows = x.shape[0]
 
     if mode == "MAP":
+        assert (
+            x.ndim == 3 and y.ndim == 3
+        ), "MAP mode requires channel dims on both x and y"
+        assert x.shape[1] == y.shape[1], "MAP mode requires channels_x == channels_y"
         shape_out = (n_rows, x.shape[1], num_out)
         out = np.zeros(shape_out, dtype=x.dtype)
         cxy = coef * x[..., idx[1]] * y[..., idx[2]]
@@ -139,7 +114,10 @@ def tensor_product_reference(
         return out.at[..., idx[0]].add(cxy)
 
     if mode == "INNER":
-        assert x.shape[-2] == y.shape[-2]
+        assert (
+            x.ndim == 3 and y.ndim == 3
+        ), "INNER mode requires channel dims on both x and y"
+        assert x.shape[1] == y.shape[1], "INNER mode requires channels_x == channels_y"
         shape_out = (n_rows, num_out)
         out = np.zeros(shape_out, dtype=x.dtype)
         cxy = coef * x[..., idx[1]] * y[..., idx[2]]
@@ -147,321 +125,16 @@ def tensor_product_reference(
         return out.at[..., idx[0]].add(sum_cxy)
 
 
-def test_forward_tensor_product():
-    """Check forward pass."""
-    num_out, idx, val, x, y = generate_tp_data()
-    params = Params(num_out=num_out, mode="OUTER")
-    expect = tensor_product_reference(idx, val, x, y, num_out)
-    result = tensor_product(pack_coef(val, idx), x, y, params)
-    assert np.allclose(expect, result, atol=1e-5)
-
-
-def test_forward_tensor_product_inner():
-    """Check forward pass."""
-    num_out, idx, val, x, y = generate_tp_data(channels=[NUM_CHANNELS] * 2)
-    params = Params(num_out=num_out, mode="INNER")
-    expect = tensor_product_reference(idx, val, x, y, num_out, mode="INNER")
-    result = tensor_product(pack_coef(val, idx), x, y, params)
-    assert np.allclose(expect, result, atol=1e-5)
-
-
-def test_jit_tensor_product():
-    """Check that e3j.ops.tensor_product can be compiled."""
-    num_out, idx, val, x, y = generate_tp_data()
-    params = Params(num_out=num_out, mode="OUTER")
-    coef = pack_coef(val, idx)
-
-    @jax.jit
-    def tp(x, y, num_out):
-        return tensor_product(coef, x, y, params)
-
-    assert tp(x, y, num_out).shape[-1] == num_out
-
-
-def test_jit_tensor_product_vmap():
-    """Check that e3j.ops.tensor_product can be compiled."""
-    num_out, idx, val, x, y = generate_tp_data()
-    params = Params(num_out=num_out, mode="OUTER")
-    coef = pack_coef(val, idx)
-    xs = np.stack([x, x])
-    ys = np.stack([y, y])
-
-    def tp(x, y):
-        return tensor_product(coef, x, y, params)
-
-    out_normal = tp(x, y)
-    out_vmap = jax.vmap(tp)(xs, ys)
-    assert out_normal.shape[-1] == num_out
-    assert out_vmap.shape[-1] == num_out
-    assert out_vmap[0].shape == out_normal.shape
-    assert np.allclose(out_vmap[0], out_normal, atol=1e-5)
-    assert np.allclose(out_vmap[1], out_normal, atol=1e-5)
-
-
-# Target config to simulate H100 GPU for testing
-h100_gpu_target_config = """
-gpu_device_info {
-  threads_per_block_limit: 1024
-  threads_per_warp: 32
-  shared_memory_per_block: 49152
-  shared_memory_per_core: 233472
-  threads_per_core_limit: 2048
-  core_count: 132
-  fpus_per_core: 128
-  block_dim_limit_x: 2147483647
-  block_dim_limit_y: 65535
-  block_dim_limit_z: 65535
-  memory_bandwidth: 3352320000000
-  l2_cache_size: 52428800
-  clock_rate_ghz: 1.98
-  device_memory_size: 84929347584
-  shared_memory_per_block_optin: 232448
-  cuda_compute_capability {
-    major: 9
-  }
-  registers_per_core_limit: 65536
-  registers_per_block_limit: 65536
-}
-runtime_version: {
-  major: 12
-  minor: 8
-  patch: 0
-}
-platform_name: "CUDA"
-dnn_version_info {
-  major: 9
-  minor: 7
-}
-device_description_str: "NVIDIA H100 80GB HBM3"
-"""
-
-
-def test_vmap_fwd_tensor_product_multi_devices():
-    """Real jax.vmap + sharding: leading vmap axis is split across 2 devices,
-    and the custom_vmap rule flattens (axis_size, num_rows) into a single
-    batch dim seen by the ffi.
-    """
-    num_out, idx, val, x, y = generate_tp_data()
-    coef = pack_coef(val, idx)
-    params = Params(num_out=num_out, mode="OUTER")
-
-    topology = get_topology_desc(
-        "name",
-        "cuda",
-        target_config=h100_gpu_target_config,
-        topology="1x1x2",
-    )
-    mesh = jax.sharding.Mesh(topology.devices, axis_names=["batch"])
-    sharding = jax.sharding.NamedSharding(
-        mesh=mesh,
-        spec=jax.sharding.PartitionSpec("batch", None, None),
-    )
-
-    def tp(x, y):
-        return tensor_product(coef, x, y, params)
-
-    tp_jitted = jax.jit(
-        jax.vmap(tp),
-        in_shardings=(sharding, sharding),
-        out_shardings=sharding,
-    )
-    tp_compiled = tp_jitted.lower(
-        jax.core.ShapedArray((128, *x.shape), x.dtype),
-        jax.core.ShapedArray((128, *y.shape), y.dtype),
-    ).compile()
-    tp_compiled_string = tp_compiled.as_text()
-
-    # What we expect to see in the HLO on each shard:
-    #   - The vmapped batch dimension (size 128) is sharded over 2 devices,
-    #     so each shard handles 64 samples. vmap then folds those into the
-    #     kernel's own row axis (NUM_ROWS=128), giving 64 * 128 = 8192 rows
-    #     fed into the forward kernel.
-    #   - Exactly one `tensor_product` custom call per shard, with inputs
-    #     (coef, x, y) and output z; coef is replicated across shards.
-    #
-    # Dimension legend (also referenced by the backward test below):
-    #   8192  = rows_per_shard = 64 (sharded vmap) * NUM_ROWS (128)
-    #   20    = x.shape[-1] = out.shape[-1] (from generate_tp_data)
-    #   30    = y.shape[-1] (from generate_tp_data)
-    #   60    = #non-zeros in coef = NUM_STRIPS (10) * LEN_STRIPS (6)
-    #   4     = packed Coef entry (1 value + 3 index components)
-    regex = re.compile(
-        r"= f32\[8192,20\]\{1,0\}.+"  # output: z
-        r'custom_call_target="tensor_product".+'
-        r"operand_layout_constraints=\{"
-        r"s32\[60,4\]\{1,0\}, "  # coef
-        r"f32\[8192,20\]\{1,0\}, "  # x
-        r"f32\[8192,30\]\{1,0\}\}",  # y
-        flags=re.MULTILINE,
-    )
-    assert len(list(regex.finditer(tp_compiled_string))) == 1, (
-        f"Expected exactly one tensor_product custom call, found "
-        f"{len(list(regex.finditer(tp_compiled_string)))}.\n\n"
-        f"HLO:\n{tp_compiled_string}"
-    )
-
-
-def test_vmap_grad_tensor_product_multi_devices():
-    """Check that the operation is sharded as expected when vmapped and jitted over 2 device.
-    This test use precompilation so it does not require a GPU to run but need to have jax version
-    with gpu support.
-
-    TODO: generalize this setup so we can use it for other tests and run them in the CI.
-    """
-    num_out, idx, val, x, y = generate_tp_data(channels=(128, None))
-    coef = pack_coef(val, idx)
-
-    # Use TRAILING_CHANNELS so the VJP routes through the tensor_product_bwd
-    params = Params(num_out=num_out, mode="OUTER", layout="TRAILING_CHANNELS")
-    x_shape = (x.shape[0], x.shape[2], x.shape[1])
-
-    topology = get_topology_desc(
-        "name",
-        "cuda",
-        target_config=h100_gpu_target_config,
-        topology="1x1x2",
-    )
-    mesh = jax.sharding.Mesh(topology.devices, axis_names=["batch"])
-    sharding = jax.sharding.NamedSharding(
-        mesh=mesh,
-        spec=jax.sharding.PartitionSpec("batch", None, None),
-    )
-    y_sharding = jax.sharding.NamedSharding(
-        mesh=mesh,
-        spec=jax.sharding.PartitionSpec("batch", None),
-    )
-
-    def tp(x, y):
-        return tensor_product(coef, x, y, params)
-
-    def sum_vmap_tp(x, y):
-        return np.sum(jax.vmap(tp)(x, y))
-
-    # A function with forward + backward tensor product passes
-    fn = jax.grad(sum_vmap_tp, argnums=(0, 1))
-
-    jitted_fn = jax.jit(
-        fn,
-        in_shardings=(sharding, y_sharding),
-        out_shardings=(sharding, y_sharding),
-    )
-    compiled_fn = jitted_fn.lower(
-        jax.core.ShapedArray((128, *x_shape), x.dtype),
-        jax.core.ShapedArray((128, *y.shape), y.dtype),
-    ).compile()  # Check that it compiles without error
-
-    compiled_fn_string = compiled_fn.as_text()
-
-    # Same shard / batch reasoning as the forward test
-    # (`test_vmap_fwd_tensor_product_multi_devices`); see the dimension
-    # legend there for 8192, 20, 30, 60, 4. Extras that only appear in the
-    # backward HLO:
-    #   128 = x channels (TRAILING_CHANNELS layout); y has no channel dim
-    #   2   = the two transposed COO orderings stacked in coef_bwd (xzy, yzx)
-    #
-    # Exactly one `tensor_product_bwd` custom call per shard, with inputs
-    # (coef_bwd, x, y, ct_z) and outputs (ct_x, ct_y).
-    regex = re.compile(
-        # outputs: (ct_x, ct_y)
-        r"\(f32\[8192,20,128\]\{2,1,0\}, f32\[8192,30\]\{1,0\}\) custom-call.+"
-        r'custom_call_target="tensor_product_bwd".+'
-        r"operand_layout_constraints=\{"
-        r"s32\[2,60,4\]\{2,1,0\}, "  # coef_bwd
-        r"f32\[8192,20,128\]\{2,1,0\}, "  # x
-        r"f32\[8192,30\]\{1,0\}, "  # y
-        r"f32\[8192,20,128\]\{2,1,0\}\}",  # ct_z
-        flags=re.MULTILINE,
-    )
-    assert len(list(regex.finditer(compiled_fn_string))) == 1, (
-        f"Expected exactly one tensor_product_bwd custom call, found "
-        f"{len(list(regex.finditer(compiled_fn_string)))}.\n\n"
-        f"HLO:\n{compiled_fn_string}"
-    )
-
-
-def test_vmap_raises_on_batched_coef():
-    num_out, idx, val, x, y = generate_tp_data()
-    coef = pack_coef(val, idx)
-    params = Params(num_out=num_out, mode="OUTER")
-    coef_batch = np.stack([coef, coef, coef])
-    with pytest.raises(jax.errors.UnexpectedTracerError):
-        jax.vmap(lambda c: tensor_product(c, x, y, params))(coef_batch)
-
-
-def test_vmap_batches_on_x_only():
-    num_out, idx, val, x, y = generate_tp_data()
-    coef = pack_coef(val, idx)
-    params = Params(num_out=num_out, mode="OUTER")
-    xs = np.stack([x, x, x])
-    zs = jax.vmap(lambda xi: tensor_product(coef, xi, y, params))(xs)
-    assert zs.shape[0] == xs.shape[0]
-    testing.assert_allclose(zs[0], zs[-1])
-
-
-def test_vmap_batches_on_y_only():
-    num_out, idx, val, x, y = generate_tp_data()
-    coef = pack_coef(val, idx)
-    params = Params(num_out=num_out, mode="OUTER")
-    ys = np.stack([y, y, y])
-    zs = jax.vmap(lambda yi: tensor_product(coef, x, yi, params))(ys)
-    assert zs.shape[0] == ys.shape[0]
-    testing.assert_allclose(zs[0], zs[-1])
-
-
-def test_backward_tensor_product():
-    """Check backward pass on x and y gradients."""
-    num_out, idx, val, x, y = generate_tp_data()
-    params = Params(num_out=num_out, mode="OUTER")
-    coef = pack_coef(val, idx)
-
-    def f_custom(x, y):
-        return np.sum(tensor_product(coef, x, y, params))
-
-    def f_ref(x, y):
-        return np.sum(tensor_product_reference(idx, val, x, y, num_out=num_out))
-
-    grads_custom = jax.grad(f_custom, argnums=(0, 1))(x, y)
-    grads_ref = jax.grad(f_ref, argnums=(0, 1))(x, y)
-
-    ct_x, ct_y = grads_custom
-    ct_x_ref, ct_y_ref = grads_ref
-    assert np.allclose(ct_x, ct_x_ref)
-    assert np.allclose(ct_y, ct_y_ref)
-
-
-def test_backward2_tensor_product():
-    """Check two TP backward passes on x and y."""
-    num_out, idx, val, x, y = generate_tp_data()
-    params = Params(num_out=num_out, mode="OUTER")
-    coef = pack_coef(val, idx)
-
-    def prod_custom(x, y):
-        return np.sum(tensor_product(coef, x, y, params))
-
-    def prod_ref(x, y):
-        return np.sum(tensor_product_reference(idx, val, x, y, num_out=num_out))
-
-    def force_grads(prod: callable) -> callable:
-        force_norm_x = lambda x, y: np.sum((jax.grad(prod)(x, y)) ** 2)
-        return jax.grad(force_norm_x, argnums=(0, 1))
-
-    d2x, d2y = force_grads(prod_custom)(x, y)
-    d2x_ref, d2y_ref = force_grads(prod_ref)(x, y)
-
-    assert np.allclose(d2x, d2x_ref)
-    assert np.allclose(d2y, d2y_ref)
-
-
 class _TestTensorProductOp:
     mode: str
-    num_idx: int
-    num_x: int
-    num_y: int
+    num_idx: int = 530
+    num_x: int = 42
+    num_y: int = 25
     num_out: int
     channels_x: int | None = None
     channels_y: int | None = None
-    num_rows = 16
-    layout: str = "LEADING_CHANNELS"
+    num_rows = 100
+    layout: str = "TRAILING_CHANNELS"
 
     def closure(self):
         """Return coefficient buffers and tensor product descriptor.
@@ -585,105 +258,107 @@ class _TestTensorProductOp:
         x, y = self.inputs()
         expect = self.fwd_ref(x, y)
         result = self.fwd_op(x, y)
-        assert_allclose(expect, result, rtol=1e-5, atol=1e-5)
+        assert_allclose(expect, result, atol=5e-5, rtol=5e-5)
 
     def test_backward(self):
         x, y = self.inputs()
         expect_dx, expect_dy = self.bwd_ref(x, y)
         result_dx, result_dy = self.bwd_op(x, y)
-        assert_allclose(expect_dx, result_dx, atol=5e-5, rtol=5e-5)
-        assert_allclose(expect_dy, result_dy, atol=5e-5, rtol=5e-5)
+        assert_allclose(expect_dx, result_dx, atol=1e-4, rtol=2e-3)
+        assert_allclose(expect_dy, result_dy, atol=1e-4, rtol=2e-3)
 
     def test_backward_backward(self):
         x, y = self.inputs()
         dz = self.fwd_ref(x, y)
         expect_Dx, expect_Dy, expect_Ddz = self.bwd_bwd_ref(x, y, dz)
         result_Dx, result_Dy, result_Ddz = self.bwd_bwd_op(x, y, dz)
-        assert_allclose(expect_Dx, result_Dx, atol=5e-5, rtol=5e-5)
-        assert_allclose(expect_Dy, result_Dy, atol=5e-5, rtol=5e-5)
-        assert_allclose(expect_Ddz, result_Ddz, atol=5e-5, rtol=5e-5)
+        assert_allclose(expect_Dx, result_Dx, atol=1e-4, rtol=2e-3)
+        assert_allclose(expect_Dy, result_Dy, atol=1e-4, rtol=2e-3)
+        assert_allclose(expect_Ddz, result_Ddz, atol=1e-4, rtol=2e-3)
 
 
+@pytest.mark.usefixtures("channels")
 class TestTensorProductOuter(_TestTensorProductOp):
+    layout = "TRAILING_CHANNELS"
     mode = "OUTER"
-    num_idx = 530
     num_x = 16
     num_y = 25
-    num_out = 12
-    channels_x = 8
-    channels_y = None
-    num_rows = 6
+    num_out = 93
+    num_rows = 100
 
 
+@pytest.mark.usefixtures("channels")
 class TestTensorProductInner(_TestTensorProductOp):
+    layout = "TRAILING_CHANNELS"
     mode = "INNER"
-    num_idx = 535
-    num_x = 12
-    num_y = 43
-    num_out = 31
-    channels_x = 4
-    channels_y = 4
-    num_rows = 12
+    num_x = 16
+    num_y = 25
+    num_out = 93
+    num_rows = 100
 
 
-class TestTensorProductOuterTrailing(TestTensorProductOuter):
+@pytest.mark.usefixtures("channels")
+class TestTensorProductMap(_TestTensorProductOp):
     layout = "TRAILING_CHANNELS"
-    # FIXME: seems to require num_channels multiple of 32
-    channels_x = 32
-    channels_y = None
-    num_rows = 22
-    num_idx = 320
+    mode = "MAP"
+    num_x = 16
+    num_y = 25
+    num_out = 93
+    num_rows = 100
 
 
-class TestTensorProductOuterTrailingCY(TestTensorProductOuterTrailing):
-    layout = "TRAILING_CHANNELS"
-    channels_x = 32
-    channels_y = None
-    num_rows = 22
-    num_idx = 320
-
-
-class TestTensorProductInnerTrailing(TestTensorProductInner):
-    layout = "TRAILING_CHANNELS"
-    channels_x = 32
-    channels_y = 32
-
-
-class TestTensorProductOuterOneIn(TestTensorProductOuter):
+@pytest.mark.usefixtures("channels")
+class TestTensorProductOuterOneIn(_TestTensorProductOp):
     layout = "TRAILING_CHANNELS"
     mode = "OUTER"
-    channels_x = 32
-    channels_y = None
     num_x = 16
     num_y = 1
     num_out = 16
     num_idx = 30
-    num_rows = 2
+    num_rows = 64
 
 
-class TestTensorProductInnerOneOut(TestTensorProductInner):
+class TestTensorProductInnerOneOut(_TestTensorProductOp):
     layout = "TRAILING_CHANNELS"
     mode = "INNER"
-    channels_x = 32
-    channels_y = 32
-    num_x = 16
-    num_y = 16
+    num_x = 25
+    num_y = 25
+    channels_x = 512
+    channels_y = 512
     num_out = 1
-    num_idx = 28
-    num_rows = 2
+    num_idx = 42
+    num_rows = 8
 
 
-class TestTensorProductMapTrailing(_TestTensorProductOp):
+class TestTensorProductOuterCY(_TestTensorProductOp):
     layout = "TRAILING_CHANNELS"
-    mode = "MAP"
-    channels_x = 32
-    channels_y = 32
+    mode = "OUTER"
+    channels_x = None
+    channels_y = 128
+    num_rows = 22
+    num_idx = 512
+    num_out = 210
+
+
+class TestTensorProductLeadingOuter(_TestTensorProductOp):
+    layout = "LEADING_CHANNELS"
+    mode = "OUTER"
+    num_idx = 780
     num_x = 16
-    num_y = 18
-    num_out = 12
-    num_idx = 132
-    num_rows = 4
+    num_y = 25
+    channels_x = 256
+    channels_y = None
+    num_out = 93
+    num_rows = 100
 
 
-if __name__ == "__main__":
-    num_out, idx, coef, x, y = generate_tp_data()
+class TestTensorProductLeadingInner(_TestTensorProductOp):
+    layout = "LEADING_CHANNELS"
+    mode = "INNER"
+    num_idx = 780
+    num_x = 16
+    num_y = 25
+    channels_x = 256
+    channels_y = 256
+    num_out = 93
+    num_rows = 100
